server/requirements.txt
click==8.1.3
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
python-decouple==3.7
Werkzeug==2.2.2
click==8.1.3
Flask==2.2.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-decouple==3.7
SQLAlchemy==2.0.0
typing_extensions==4.4.0
Werkzeug==2.2.2
click==8.1.3
Flask==2.2.2
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2-binary==2.9.5
python-decouple==3.7
SQLAlchemy==2.0.0
typing_extensions==4.4.0
Werkzeug==2.2.2
